pyproject.toml
[project]
name = "serka"
version = "0.1.0"
description = "Serka (serĉi ekologio) is a protoype search tool for the UKCEH NCUK project. It uses modern AI techniques (LLMs, RAG, Vector Stores) to enhance the EIDC catalogue's search functionality."
readme = "README.md"
authors = [
    { name = "Matthew Coole", email = "matcoo@ceh.ac.uk" }
]
requires-python = ">=3.12"
dependencies = [
    "chromadb>=0.6.1",
    "fastapi[standard]>=0.115.6",
    "haystack-ai>=2.9.0",
    "jq>=1.8.0",
    "ollama-haystack>=2.2.0",
    "ollama>=0.4.5",
    "python-dotenv>=1.0.1",
    "chroma-haystack>=2.0.0",
    "pyyaml>=6.0.2",
    "trafilatura>=2.0.0",
    "beautifulsoup4>=4.13.0",
    "scrapy>=2.12.0",
    "pymongo>=4.11.1",
    "neomodel>=5.4.5",
    "neo4j-haystack>=2.2.1",
]

[project.scripts]
serka = "serka:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ipython>=8.31.0",
    "mypy>=1.14.1",
    "pre-commit>=4.0.1",
    "pytest>=8.3.4",
    "ruff>=0.8.6",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "tab"

[tool.mypy]
files = ["src", "tests"]
strict = true

[tool.pytest.ini_options]
markers = [
    "integration: Marks tests as integration tests that may be slow.",
    "unit: Marks tests as unit tests that should be fast.",
]
